Fix last checkstyle failures
