+++ Competition2018/src/test/java/competition/subsystems/lean/commands/LeanRightCommandTest.java
@@ -1,23 +1,23 @@
-package competition.subsystems.lean.Commands;
+package competition.subsystems.lean.commands;
 
 import static org.junit.Assert.assertEquals;
 import org.junit.Test;
 
 import competition.subsystems.lean.LeanSubsystem;
-import competition.subsystems.lean.commands.IncreaseSpeedCommand;
+import competition.subsystems.lean.commands.LeanRightCommand;
 import xbot.common.injection.BaseWPITest;
 
-public class IncreaseSpeedCommandTest extends BaseWPITest {
+public class LeanRightCommandTest extends BaseWPITest {
 
     LeanSubsystem lean;
-    IncreaseSpeedCommand command;
+    LeanRightCommand command;
 
     @Override
     public void setUp() {
         super.setUp();
 
         lean = injector.getInstance(LeanSubsystem.class);
-        command = injector.getInstance(IncreaseSpeedCommand.class);
+        command = injector.getInstance(LeanRightCommand.class);
 
         lean.temporaryHack();
     }
@@ -29,13 +29,10 @@ public class IncreaseSpeedCommandTest extends BaseWPITest {
     }
 
     @Test
-    public void checkIncreaseSpeed() {
-        lean.decreaseSpeed();
-        lean.leanRight();
-        assertEquals(-0.1, lean.motor.getMotorOutputPercent(), 0.001);
+    public void checkLeanRight() {
+        assertEquals(0.0, lean.motor.getMotorOutputPercent(), 0.001);
         command.initialize();
         command.execute();
-        lean.leanRight();
         assertEquals(-0.4, lean.motor.getMotorOutputPercent(), 0.001);
     }
 }
